pyproject.toml
[project]
name = "reddit-rag-analytics-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-snowflake>=1.11.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.29.0",
    "pandas>=3.0.1",
    "praw>=7.8.1",
    "pyarrow>=23.0.1",
    "pyspark==3.5.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "snowflake-connector-python>=4.3.0",
    "tiktoken>=0.12.0",
]
